Added query to get features
To get all the features in one of the album XMLs, open this file in
oxygen, hit the icon in the top right corner that has an XQ, and then
click the plain blue arrow. It will spit out all of the features in
order, with one extra pipe at the end that you’ll have to delete
(you’ll get an error). This’ll show if we have any misspellings/repeats
that need fixed. NOTE this will only work after all of the album xmls
match Paige’s/the current schema.